fix: rewrite of EmperorsManager
The new code is much cleaner and safer.  It also keeps a list of old
emperors for when the last position becomes vacant.  Clearing the
leaderboard or removing a single emperor with /setemperors now
triggers the correct events.  This rewrite was actually done in Rust
and thoroughly tested before being translated back to Java.